+++ test/main.cpp
@@ -1,6 +1,7 @@
 #include <fstream>
 #include <iostream>
 
+#include "../album.h"
 #include "../cue.h"
 #include <nlohmann/json.hpp>
 
@@ -12,6 +13,6 @@ int main(int argc, char** argv) {
   std::ifstream f(argv[1], std::ios::in);
   nlohmann::json j;
   f >> j;
-  generate(j, argv[2]);
+  cue::generate(Album::from_json(j), argv[2]);
   return 0;
 }
